requirements.txt
asgiref==3.8.1
boto3==1.40.36
botocore==1.40.36
Django==5.2.3
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-rest-authtoken==2.1.4
django-storages==1.14.6
djangorestframework==3.16.1
Faker==37.8.0
faker-commerce==1.0.4
jmespath==1.0.1
Markdown==3.9
mysqlclient==2.2.7
pillow==11.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
